Deprecate CMakePropertiesController
CMakePropertiesController was only used to create a new
CMakeProperties.  The load() method loads a YAML file
.settings/CDT-cmake.yaml which never existed, because the save()
method was never called.  There was no separate YAML file for Run and
Debug.

CMake properties are stored in the project preferences by the launch
config UI and these properties are used to construct the cmake command
line.