Adds capability to display warnings in LaunchBar Launch Config Dialog
Add SWTBot to test that the Launch Configuration, when opened via the
Launch Bar, displays a warning in the message area.

The Launch Bar Launch Configuration is, confusingly, handled by 2
different classes depending on whether a new configuration is being
created (NewLaunchConfigEditPage) or edited
(LaunchBarLaunchConfigDialog).

When using NewLaunchConfigEditPage, the existing
LaunchConfigurationTabGroupViewer.getWarningMessage() mechanism is used.
This was added to eclipse-platform in Bug 386673 (commit 231ef13).

When using LaunchBarLaunchConfigDialog (when editing), the new
getWarningMessage() mechanism, copied from the existing, is used.

In both classes above, logic was added to update the message when a tab
change occurs.